adiciona relacao pedido-funcionario

+++ back-end/gerenciador_de_pedidos/views.py
@@ -1,5 +1,7 @@
 from gerenciador_de_clientes.models import Cliente
 from gerenciador_de_clientes.serializers import ClienteSerializer
+from gerenciador_de_funcionarios.models import Funcionario
+from gerenciador_de_funcionarios.serializers import FuncionarioSerializer
 from .models import Pedido
 from .serializers import PedidoSerializerRequest, PedidoSerializerResponse
 from rest_framework import status
@@ -10,13 +12,17 @@ from rest_framework.views import APIView
 
 class PedidoView(APIView):
     def post(self, request):
-        print(request.data)
+
         serializer = PedidoSerializerRequest(data=request.data)
         serializer.is_valid(raise_exception=True)
 
         cliente_id = request.data.get('cliente', None)
-        pedido = serializer.save(cliente_id=cliente_id)
+        funcionario_id = request.data.get('funcionario', None)
 
+        if not Funcionario.objects.filter(id=funcionario_id).exists():
+            return Response("Funcionário não encontrado", status=status.HTTP_400_BAD_REQUEST)
+
+        pedido = serializer.save(cliente_id=cliente_id, funcionario_id=funcionario_id)
         response_data = PedidoSerializerResponse(pedido).data
 
         return Response(data=response_data, status=status.HTTP_201_CREATED)
@@ -25,19 +31,38 @@ class PedidoView(APIView):
 class PedidoListView(APIView):
     def get(self, request):
         pedidos = Pedido.objects.all()
-        serializer = PedidoSerializerResponse(pedidos, many=True)
-        pedidos_data = serializer.data
+        pedidos_data = []
 
-        for pedido in pedidos_data:
-            cliente_id = pedido['cliente']
+
+        for pedido in pedidos:
+            pedido_data = PedidoSerializerResponse(pedido).data
+
+            # Adiciona informações sobre o cliente
+            cliente_id = pedido_data['cliente']
             try:
                 cliente = Cliente.objects.get(pk=cliente_id)
-                pedido['cliente'] = ClienteSerializer(cliente).data
+                pedido_data['cliente'] = ClienteSerializer(cliente).data
             except Cliente.DoesNotExist:
-                pedido['cliente'] = None
+                pedido_data['cliente'] = None
 
-            total = sum(float(produto['preco']) for produto in pedido['produtos'])
-            pedido['total'] = total
+            # Adiciona informações sobre o funcionário
+            funcionario_id = pedido.funcionario_id
+            if funcionario_id:
+                try:
+                    funcionario = Funcionario.objects.get(id=funcionario_id)
+                    pedido_data['funcionario'] = {
+                        'id': funcionario.id,
+                        'nome': funcionario.nome
+                    }
+                except Funcionario.DoesNotExist:
+                    pedido_data['funcionario'] = None
+            else:
+                pedido_data['funcionario'] = None
+
+        total = sum(float(produto['preco']) for produto in pedido_data['produtos'])
+        pedido_data['total'] = total
+
+        pedidos_data.append(pedido_data)
 
         return Response(data=pedidos_data, status=status.HTTP_200_OK)
 
@@ -62,6 +87,15 @@ class PedidoDetailView(APIView):
         cliente_data = ClienteSerializer(cliente).data
         serializer.data['cliente'] = cliente_data
 
+        funcionario_id = pedido.funcionario_id
+        if funcionario_id:
+            funcionario = Funcionario.objects.get(pk=funcionario_id)
+            funcionario_data = FuncionarioSerializer(funcionario).data
+            serializer.data['funcionario'] = funcionario_data
+        else:
+            serializer.data['funcionario'] = None
+
+
         return Response(data=data)
 
 
@@ -74,8 +108,12 @@ class PedidoUpdateView(APIView):
 
     def put(self, request, pk):
         pedido = self.get_object(pk)
-        serializer = PedidoSerializerRequest(instance=pedido, data=request.data, partial=True)
-        print(request.data)
+        data = request.data.copy()
+
+        if 'funcionario' not in data:
+            return Response({"error": "O campo 'funcionario' é obrigatório"}, status=400)
+
+        serializer = PedidoSerializerRequest(instance=pedido, data=data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response("Atualizado com sucesso")
@@ -90,5 +128,8 @@ class PedidoDeleteView(APIView):
 
     def delete(self, request, pk):
         pedido = self.get_object(pk)
+        if pedido.funcionario is None:
+            return Response({"error": "Não é possível excluir o pedido pois não está associado a um funcionário"},
+                            status=status.HTTP_400_BAD_REQUEST)
         pedido.delete()
         return Response(status=status.HTTP_202_ACCEPTED, data="Pedido deletado com sucesso")
